Update 0.1
I just started to use some grids and media in CSS to make it responsive. Just making around with it rn and learning some things I missed. Still new to responsive design so this will project will help me out a lot and will learn a lot at the end of it. This is just the basic nav bar that's not close to the finish but I am just playing around with the px and rows rn. The next thing I am going to work on is making it look clean and also making a clean response design to  a table and phone size px screen.