changes 0.2
Done with the nav bar and about/home sections. Adding the skills section right now and then going to move on with projects and about me which should take a lot longer. After finishing the base of the portfolio, I will add some CSS to make everything look nice and clean with some animations and hover effects. The last Task would be to add some javascript when you click the button on the top it will auto-scroll you down to that section on the page. This will be my first time messing around with something like this so it should be exciting and hopefully not too complex. I also am thinking about adding some animation when you scroll the text will come from the side of the screen. Everything on the page looks excellent from sizing down the window screen so far. Going to add and sub some padding once it's at an iPhone px but rn just focusing on finishing the basic stuff first. This was around 2 hours' worth of work from last time.